Send Track Aux To A Mixbus Track
For ex: If you have 20 tracks and you want to send all 20 or just 10 tracks to a track called MIXBUS (You can name it what ever you want in the script)

At the bottom you have a track called "MIXBUS"

Just select what ever tracks you want and hit this action. Then it will remove the "MASTER SEND" and create an AUX to the track "MIXBUS".